fix(web): key semantic status validators

+++ mammamiradio/web/streamer.py
@@ -10906,7 +10906,7 @@ def _public_status_payload(request: Request, *, include_etag_revision: bool = Fa
     # the segment now_streaming is playing (send-start is provisional).
     recent_moments: list[dict] = []
     recent_revision: tuple[str, ...] = ()
-    event_revision: int | None = None
+    event_revision: float | None = None
     ha_capable = bool(config.ha_token and config.homeassistant.enabled and config.homeassistant.context_enabled)
     moment_store = getattr(state, "moment_store", None)
     authorization = state.home_authorization or HomeAuthorization.narrow()
@@ -10934,7 +10934,7 @@ def _public_status_payload(request: Request, *, include_etag_revision: bool = Fa
         if state.ha_last_event_ts > 0 and (_now - state.ha_last_event_ts) < _retention:
             ha_moments["last_event_label"] = state.ha_last_event_label
             ha_moments["last_event_ago_min"] = max(1, round((_now - state.ha_last_event_ts) / 60))
-            event_revision = int(state.ha_last_event_ts // 60)
+            event_revision = state.ha_last_event_ts
         if state.ha_ritual_public_families:
             ha_moments["ritual_families"] = list(state.ha_ritual_public_families[:4])
         if recent_moments:
